FIX: Only one (XML / HTML) tag is swapped.
This was broken by the switch from register s to the default register in 4e667c248a2f469dcfb065c47a5f5c07075149b7, because the visual mode paste clobbers that register (with the replaced contents, which here is the previous tag name).
Split the command into two parts, and restore the default register in between.