Check for active matchit.vim plugin, and else skip XML tag swap.
The commands for XML tag swapping require the modified "%" command provided by the matchit.vim plugin. Without it, the commands will wreak havoc on the text.
Add a check for the plugin and if not installed, fall back to the general swap, which will only work on the current tag (not the tag pair).